fuature - added folder with handlers:
- added folder with handlers in categories app;
- added new schema (BadRequestSchema) in base folder;
- added new constants in settings.constants.
