+++ src/apps/categories/handlers/__init__.py
@@ -1,0 +1,2 @@
+from .category_handler import CategoryHandler
+from .categories_handler import CategoriesHandler
