+++ src/settings/constants.py
@@ -3,6 +3,7 @@ HTTP_200_OK = 200
 HTTP_201_CREATED = 201
 HTTP_204_NO_CONTENT = 204
 
+HTTP_400_BAD_REQUEST = 400
 HTTP_404_NOT_FOUND = 404
 
 # BODY FOR ERRORS
